requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
httpx
python-multipart